feat(react): make Radio component
