+++ docs/app/react/radio/components/radio-preview.tsx
@@ -1,37 +1,5 @@
-import { focusStates } from '@cerberus-design/panda-preset'
-import { Field, Label } from '@cerberus-design/react'
-import { css, cx } from '@cerberus/styled-system/css'
+import { Field, Label, Radio } from '@cerberus-design/react'
 import { hstack } from '@cerberus/styled-system/patterns'
-import { radio } from '@cerberus/styled-system/recipes'
-import { type InputHTMLAttributes, type PropsWithChildren } from 'react'
-
-interface RadioProps
-  extends Omit<InputHTMLAttributes<HTMLInputElement>, 'size'> {
-  id: string
-  size?: 'sm' | 'md'
-}
-
-export function Radio(props: PropsWithChildren<RadioProps>) {
-  const { children, size, ...nativeProps } = props
-  const styles = radio({
-    size,
-  })
-  return (
-    <Field>
-      <div className={cx('group', hstack(), styles.root)} tabIndex={0}>
-        <input
-          {...nativeProps}
-          className={styles.input}
-          tabIndex={-1}
-          type="radio"
-        />
-        <Label className={styles.label} htmlFor={props.id}>
-          {children}
-        </Label>
-      </div>
-    </Field>
-  )
-}
 
 export function OverviewRadioGroup() {
   return (
@@ -44,15 +12,23 @@ export function OverviewRadioGroup() {
       name="pet"
       role="radiogroup"
     >
-      <Radio id="pet" name="pet" value="dog" defaultChecked>
-        🐶 Dog
-      </Radio>
-      <Radio id="pet" name="pet" value="cat">
-        🐱 Cat
-      </Radio>
-      <Radio id="pet" name="pet" value="mouse">
-        🐭 Mouse
-      </Radio>
+      <Field>
+        <Radio id="dog" name="pet" value="dog" defaultChecked>
+          <Label htmlFor="dog">🐶 Dog</Label>
+        </Radio>
+      </Field>
+
+      <Field>
+        <Radio id="cat" name="pet" value="cat">
+          <Label htmlFor="cat">😸 Cat</Label>
+        </Radio>
+      </Field>
+
+      <Field>
+        <Radio id="both" name="pet" value="both">
+          <Label htmlFor="both">🐶😸 Both</Label>
+        </Radio>
+      </Field>
     </fieldset>
   )
 }
@@ -68,12 +44,16 @@ export function OverviewRadioSizes() {
       name="sizes"
       role="radiogroup"
     >
-      <Radio id="sizes" name="sizes" value="sm" size="sm">
-        Small
-      </Radio>
-      <Radio id="sizes" name="sizes" value="md">
-        Medium (default)
-      </Radio>
+      <Field>
+        <Radio id="sm" name="sizes" value="sm" size="sm">
+          <Label htmlFor="sm">Small</Label>
+        </Radio>
+      </Field>
+      <Field>
+        <Radio id="md" name="sizes" value="md">
+          <Label htmlFor="md">Medium (default)</Label>
+        </Radio>
+      </Field>
     </fieldset>
   )
 }
