+++ packages/react/src/index.ts
@@ -13,6 +13,7 @@ export * from './components/Label'
 export * from './components/NavMenuTrigger'
 export * from './components/NavMenuList'
 export * from './components/NavMenuLink'
+export * from './components/Radio'
 export * from './components/Tab'
 export * from './components/TabList'
 export * from './components/TabPanel'
